DBD-SQLite: now BegunWork (set by begin_work) is handled properly (as requested in DBI::Changes; see notes for 1.20), and this resolved rt #48393.
In fact, it looks like I lost this code by mistake while preparing the previous release. I remember I wrote this before...

I added a test for this, but as it requires the perl debugger, it is disabled by default. Wondering if we can do it without calling system() and the likes...