WiX.pm: Fix bug in output_base_filename
Fragment/Files.pm: Make the file ID be a base64'd CRC32 instead of a GUID - hopefully there are no collisions. 

Reason is that the file tag's ID within a merge module needs to be shorter, and it's the merge module's package GUID appended to the actual ID given here.