When a subroutine attribute spans more than one line, the PPI::Token::Attribute contains the first line of the attribute, and as many new line characters as there are lines remaining in the file. You can reproduce this behavior with something like
sub foo : Bar(
bazzle);

This appears to be happening in sub __TOKENIZER__scan_for_end(), which
calls $t->_fill_line(1) to get subsequent lines, and then goes back to
the top of the while() loop to figure out whether the end of the
attribute is found. The construction of the $search string seems to
assume that $t->{line_cursor} will be 0 at this point, but
PPI::Tokenizer->_fill_line() sets it to -1 (assuming, I believe, that it
will be pre-incremented by that package's _process_next_char() before
use). The substr() built-in takes this as a request to start that many
characters from the end of the line, and the observed behavior follows
from this.

Pragmatically, the minimum change appears to be to set $t->{line_cursor}
to 0 just after the call to $t->_fill_line(1), and this I have done.
This seems to produce correct results for both single-line and
multi-line attributes, but I have _not_ tried this out on entities other
than subroutines, and you are invited _not_ to take my word for the
behavior with subroutines.

I am sure there are other ways to address this behavior, and you may
prefer some other fix for architectural or editorial reasons.