Actually found in the wild: someone created a method named 'sub', and invoked it. PPI thought the invocation was a subroutine declaration, and got terribly confused.
The patch is a line in the test for an anonymous subroutine that
prevents the test from succeeding if the '->' operator precedes the
'sub'.

Fortunately, it apears that Perl fails to parse an attempt to call a
subroutine named 'sub' as 'sub()', and '&sub()' and 'main::sub()' take a
different path through the code.

This appears to me to be such a rare thing that I was not going to file
an RT ticket to go with the patch, though I will if you like.

Tom (wyant at cpan dot org)